Fix an overflow bug in pawns stormValue
These fields are defined as int8_t but values bigger
then 127 are stored there so that we silently overflow.

Fix bringing up all the fields to a sane int type. This
will increase memory usage, but apart from being safe, it is
not clear if code is slower or faster. Test is needed.

+++ src/pawns.h
@@ -58,9 +58,9 @@ private:
 
   Key key;
   Bitboard passedPawns;
-  int16_t mgValue, egValue;
-  int8_t ksStormValue[2], qsStormValue[2];
-  uint8_t halfOpenFiles[2];
+  int mgValue, egValue;
+  int ksStormValue[2], qsStormValue[2];
+  int halfOpenFiles[2];
 };
